ignite-2.4.1-shared-folder Fixed race in WAL between rollover() and allocateRemainingFiles()

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/wal/FileWriteAheadLogManager.java
@@ -1169,10 +1169,10 @@ public class FileWriteAheadLogManager extends GridCacheSharedManagerAdapter impl
         if (allFiles.length == 0) {
             File first = new File(walWorkDir, FileDescriptor.fileName(0));
 
-            createFile(first);
+            createFile(first, true);
         }
         else
-            checkFiles(0, false, null);
+            checkFiles(0, false, true, null);
     }
 
     /**
@@ -1212,7 +1212,7 @@ public class FileWriteAheadLogManager extends GridCacheSharedManagerAdapter impl
      * @param file File to create.
      * @throws IgniteCheckedException If failed.
      */
-    private void createFile(File file) throws IgniteCheckedException {
+    private void createFile(File file, boolean failIfAlreadyExists) throws IgniteCheckedException {
         if (log.isDebugEnabled())
             log.debug("Creating new file [exists=" + file.exists() + ", file=" + file.getAbsolutePath() + ']');
 
@@ -1223,13 +1223,14 @@ public class FileWriteAheadLogManager extends GridCacheSharedManagerAdapter impl
         try {
             Files.move(tmp.toPath(), file.toPath());
         }
-        catch (FileAlreadyExistsException e) {
-            // Rollover() already created segment file.
-            tmp.delete();
-
-            return;
-        }
         catch (IOException e) {
+            if (!failIfAlreadyExists && (e instanceof FileAlreadyExistsException)) {
+                // Rollover() already created segment file.
+                tmp.delete();
+
+                return;
+            }
+
             throw new IgniteCheckedException("Failed to move temp file to a regular WAL segment file: " +
                 file.getAbsolutePath(), e);
         }
@@ -1633,7 +1634,7 @@ public class FileWriteAheadLogManager extends GridCacheSharedManagerAdapter impl
          * {@link FileWriteAheadLogManager#checkOrPrepareFiles()}
          */
         private void allocateRemainingFiles() throws IgniteCheckedException {
-            checkFiles(1, true, new IgnitePredicate<Integer>() {
+            checkFiles(1, true, false, new IgnitePredicate<Integer>() {
                 @Override public boolean apply(Integer integer) {
                     return !checkStop();
                 }
@@ -1955,10 +1956,12 @@ public class FileWriteAheadLogManager extends GridCacheSharedManagerAdapter impl
      *
      * @param startWith Start with.
      * @param create Flag create file.
+     * @param failIfAlreadyExists If false, file creating conflict will be tolerated.
      * @param p Predicate Exit condition.
      * @throws IgniteCheckedException if validation or create file fail.
      */
-    private void checkFiles(int startWith, boolean create, IgnitePredicate<Integer> p) throws IgniteCheckedException {
+    private void checkFiles(int startWith, boolean create, boolean failIfAlreadyExists, IgnitePredicate<Integer> p)
+        throws IgniteCheckedException {
         for (int i = startWith; i < dsCfg.getWalSegments() && (p == null || (p != null && p.apply(i))); i++) {
             File checkFile = new File(walWorkDir, FileDescriptor.fileName(i));
 
@@ -1971,7 +1974,7 @@ public class FileWriteAheadLogManager extends GridCacheSharedManagerAdapter impl
                         "(WAL segment size change is not supported):" + checkFile.getAbsolutePath());
             }
             else if (create)
-                createFile(checkFile);
+                createFile(checkFile, failIfAlreadyExists);
         }
     }
